Add standalone remarks page and shared panel service
