+++ Program.cs
@@ -176,6 +176,7 @@ builder.Services.AddScoped<ProjectFactsReadService>();
 builder.Services.AddScoped<ProjectProcurementReadService>();
 builder.Services.AddScoped<ProjectTimelineReadService>();
 builder.Services.AddScoped<ProjectCommentService>();
+builder.Services.AddScoped<ProjectRemarksPanelService>();
 builder.Services.AddScoped<IRemarkService, RemarkService>();
 builder.Services.AddScoped<IRemarkNotificationService, RemarkNotificationService>();
 builder.Services.AddSingleton<IRemarkMetrics, RemarkMetrics>();
